refactor(server): split xmpp ffi boundary modules
